+++ QuizzApp/QuizzApp/Helpers/AppSettings.cs
@@ -1,4 +1,4 @@
-namespace WebApi.Helpers
+namespace QuizzApp.Helpers
 {
     public class AppSettings
     {
