+++ QuizzApp/QuizzApp/Helpers/AuthorizeAttribute.cs
@@ -1,8 +1,9 @@
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
+using QuizzApp.Models;
 using System;
-using WebApi.Entities;
+
 
 [AttributeUsage(AttributeTargets.Class | AttributeTargets.Method)]
 public class AuthorizeAttribute : Attribute, IAuthorizationFilter
