+++ QuizzApp/QuizzApp/Helpers/JwtMiddleware.cs
@@ -1,14 +1,15 @@
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Options;
 using Microsoft.IdentityModel.Tokens;
+using QuizzApp.Services;
 using System;
 using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using WebApi.Services;
 
-namespace WebApi.Helpers
+
+namespace QuizzApp.Helpers
 {
     public class JwtMiddleware
     {
