+++ QuizzApp/QuizzApp/Models/ApplicationContext.cs
@@ -18,5 +18,6 @@ namespace QuizzApp.Models
         public DbSet<Question> Questions { get; set; }
         public DbSet<Answer> Answers { get; set; }
         public DbSet<Result> Results { get; set; }
+        public DbSet<User> Admins { get; set; }
     }
 }
