+++ QuizzApp/QuizzApp/Services/UserService.cs
@@ -1,16 +1,16 @@
 using Microsoft.Extensions.Options;
 using Microsoft.IdentityModel.Tokens;
+using QuizzApp.Helpers;
+using QuizzApp.Models;
 using System;
 using System.Collections.Generic;
 using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
 using System.Security.Claims;
 using System.Text;
-using WebApi.Entities;
-using WebApi.Helpers;
-using WebApi.Models;
 
-namespace WebApi.Services
+
+namespace QuizzApp.Services
 {
     public interface IUserService
     {
